misc.js
